+++ R/count-tally.R
@@ -113,6 +113,26 @@ tally.tbl_sql <- tally.data.frame
 #' @export
 #' @rdname count
 add_count <- function(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = deprecated()) {
+  UseMethod("add_count")
+}
+
+#' @export
+add_count.default <- function(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = deprecated()) {
+  if (!missing(.drop)) {
+    lifecycle::deprecate_warn("1.0.0", "add_count(.drop = )")
+  }
+
+  if (!missing(...)) {
+    out <- group_by(x, ..., .add = TRUE)
+  } else {
+    out <- x
+  }
+  add_tally(out, wt = !!enquo(wt), sort = sort, name = name)
+}
+
+
+#' @export
+add_count.data.frame <- function(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = deprecated()) {
   if (!missing(.drop)) {
     lifecycle::deprecate_warn("1.0.0", "add_count(.drop = )")
   }
@@ -123,10 +143,7 @@ add_count <- function(x, ..., wt = NULL, sort = FALSE, name = NULL, .drop = depr
     out <- x
   }
   out <- add_tally(out, wt = !!enquo(wt), sort = sort, name = name)
-  if (is.data.frame(x)) {
-    out <- dplyr_reconstruct(out, x)
-  }
-  out
+  dplyr_reconstruct(out, x)
 }
 
 #' @rdname count