+++ NEWS.md
@@ -16,6 +16,8 @@
 
 * row-wise data frames of 0 rows and list columns are supported again (#5804).
 
+* `add_count()` is now generic (#5837).
+
 # dplyr 1.0.5
 
 * Fixed edge case of `slice_sample()` when `weight_by=` is used and there